Add files via upload
Finished dictionary and tradition methods. They will now generate passwords based on the checkboxes. 
Added css styling to button and output with transitions as well.

Small bug in the dictionary where with option 0 and 2 it duplicates the 'e' for whatever reason. Also working on some catches for the traditional side.